test/quant_risk_timing/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
PyYAML>=6.0.0
tushare>=1.2.89
pytest>=7.0.0
